extract matchesPattern helper
